feature: trainers index refactor and sliders
