+++ src/components/App.module.css
@@ -4,4 +4,5 @@
   top: 40px;
   border: none;
   background-color: transparent;
+  font-size: 30px;
 }
