+++ src/components/TrainersIndex/TrainerItem.js
