+++ src/index.css
@@ -15,4 +15,5 @@ ul {
 
 body {
   padding-bottom: 100px;
+  background-color: #fcfcfc;
 }
